pyproject.toml
[tool.poetry]
name = "flask-api"
version = "0.1.0"
description = "A generic flask api with auto-generated swagger docs."
authors = ["A Guy Who Needs A Distraction"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
flask-restx = "^1.3.0"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"
pymysql = "^1.1.1"
cryptography = "^44.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
flask-testing = "^0.8.1"
coverage = "^7.6.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
